Stop quantizing the SenseNova training timestep to bf16
train_step cast t to training_dtype, which train_runner forces to bf16 for
this arch. t is a VALUE the model is conditioned on, not an activation:
timestep_embedder embeds it and noise_scale_embedder derives its input from
the same tensor, so bf16 rounded the conditioning itself to ~2e-3
(0.3123456 -> 0.3125). Inference conditions on the fp32 t that linspace
produces, and the sampler upstream is fp32 too, so the cast was creating a
train/inference skew out of nothing.

Keep t fp32 and cast only where a dtype is actually required: z_image casts
explicitly so the noised image stays training_dtype, matching what
inference gets from 0-dim promotion when it noises in the image dtype -
_build_step_context's patchify and gen-branch ViT run outside the autocast
block, so letting z_image promote would have changed their input dtype.
The velocity denominator is now fp32, which lifts v_pred/v_target into the
fp32 the MSE already asked for, making the .float() calls no-ops instead of
extra copies.

The existing vendor-math parity test runs at training_dtype fp32, where the
old cast was a no-op, so it was blind to this. Adds a bf16 case that pins
the timestep reaching _build_step_context as unrounded fp32 and z_image as
training_dtype.

+++ backend/core/training/ops/sensenova_ops.py
@@ -248,7 +248,10 @@ def train_step(
             t = t[0]
     else:
         t = timesteps
-    t = torch.as_tensor(t, device=device, dtype=dtype).reshape(-1)
+    # Keep t in fp32, the dtype the sampler produces and the dtype inference's
+    # `ts` carries (linspace, sensenova_pipeline_ops.py:1068): timestep_embedder
+    # embeds t's VALUE, and bf16 would quantize it to ~2e-3 in training only.
+    t = torch.as_tensor(t, device=device, dtype=torch.float32).reshape(-1)
     if t.numel() != 1:
         raise ValueError("SenseNova training requires one timestep for batch_size=1")
 
@@ -263,7 +266,10 @@ def train_step(
         raise ValueError("SenseNova image does not align to the merged token grid")
     token_h, token_w = grid_h // merge_size, grid_w // merge_size
     noise_scale = compute_noise_scale(transformer, grid_h, grid_w, merge_size)
-    z_image = t.view(1, 1, 1, 1) * x0 + (1 - t).view(1, 1, 1, 1) * (
+    # Inference noises in the image dtype, its fp32 t demoted by 0-dim promotion
+    # (sensenova_pipeline_ops.py:1122); cast explicitly so z_image stays
+    # training_dtype -- _build_step_context's ViT runs outside the autocast below.
+    z_image = t.to(dtype).view(1, 1, 1, 1) * x0 + (1 - t).to(dtype).view(1, 1, 1, 1) * (
         torch.randn_like(x0) * noise_scale
     )
     shape = SimpleNamespace(
@@ -305,6 +311,8 @@ def train_step(
             .view(1, token_h * token_w, patch * patch * 3)
         )
         x0_tokens = transformer.patchify(x0, patch)
+        # fp32 t here lifts v into fp32, which the MSE below wanted anyway --
+        # the .float() calls become no-ops rather than extra copies.
         denominator = (1 - t).view(1, 1, 1).clamp_min(transformer.config.t_eps)
         v_pred = (x0_pred - z) / denominator
         v_target = (x0_tokens - z) / denominator
